fixed login modal

+++ components/@web/Features/@common/FavoriteButton.tsx
@@ -1,26 +1,33 @@
+import { CurrentUserContext, CurrentUserContextType } from "@/context/currentUser.context";
 import { WebPreviewContext, WebPreviewContextType } from "@/context/webPreview.context";
 import useAxiosAuth from "@/hooks/useAxiosAuth";
 import { toast } from "@/lib/toast";
-import { ReportFormData } from "@/types/formsData";
 import { Favorite, FavoriteBorderOutlined } from "@mui/icons-material";
 import { Tooltip } from "antd";
 import { useContext, useState } from "react";
 
+//
+
 const FavoriteButton = ({ isFav, size = 28 }: { isFav: boolean; size?: number }) => {
   const [isFavorite, setFavorite] = useState(isFav);
   const { relatedTo, relatedToID } = useContext(WebPreviewContext) as WebPreviewContextType;
+  const { showLoginModal, setShowLoginModal, isLogin } = useContext(CurrentUserContext) as CurrentUserContextType;
 
   const api = useAxiosAuth();
   const onSubmit = async () => {
+    if (!isLogin) {
+      setShowLoginModal(true);
+      return;
+    }
     try {
       if (isFavorite) {
-        await api.post(`/estate/favorite/${relatedToID}`);
-        toast.success("در لیست مورد علاقه ها قرار گرفت");
-        setFavorite(true);
-      } else {
         await api.delete(`/estate/favorite/${relatedToID}`);
         toast.success("از لیست حذف شد");
         setFavorite(false);
+      } else {
+        await api.post(`/estate/favorite/${relatedToID}`);
+        toast.success("در لیست مورد علاقه ها قرار گرفت");
+        setFavorite(true);
       }
     } catch (error) {
       //
